+++ .eslintrc.js
@@ -13,7 +13,5 @@ module.exports = {
   ],
   env: {
     browser: true
-  },
-  rules: {
   }
 };